We need to wait just long enough to pass the collection interval - don't wait for exactly the collection interval, we need to wait for it to expire.
1. the operation name parameter string the client gives us is the operation's name (getName) not its ID (getID) (even though the operation entity's name and ID are the same today, it doesn't have to be and could change in the future.)
2. javadoc Operation

HWKAGENT-52 - do not allow an app to be deployed unless the server has a "deploy" operation defined.
this feature is to allow a user to disable deployments by simply removing the deploy operation metadata from the server resource type (as will be the case for Swarm app server builds)

Notice I moved the "deploy" operation out of the Deployment resource type into the parent WildFly Server type.
"Deploy" operation creates new deployments - so it doesn't make sense to have a "Deploy" operation on a deployment itself
(you don't deploy an existing deployment.) You can REdeploy a deployment, you can REMOVE a deployment, but it is incoherent to say
you are deploying an existing deployment - its already deployed!
You deploy a new deployment on a server, which is why the operation makes more sense being on the server resource type.

add some more javadoc

fix license